username

+++ routes/ideas.js
@@ -44,17 +44,27 @@ router.post('/', async (req, res) => {
 //Update an idea
 router.put('/:id', async (req, res) => {
 	try {
-		const upadtedIdea = await Idea.findByIdAndUpdate(
-			req.params.id,
-			{
-				$set: {
-					text: req.body.text,
-					tag: req.body.tag,
+		const idea = await Idea.findById(req.params.id);
+
+		//Match the usernames
+		if (idea.username === req.body.username) {
+			const upadtedIdea = await Idea.findByIdAndUpdate(
+				req.params.id,
+				{
+					$set: {
+						text: req.body.text,
+						tag: req.body.tag,
+					},
 				},
-			},
-			{ new: true }
-		);
-		res.json({ success: true, data: upadtedIdea });
+				{ new: true }
+			);
+			return res.json({ success: true, data: upadtedIdea });
+		}
+
+		//Usernames don't match
+		else {
+			res.status(403).json({ success: false, error: 'Unauthorized to update' });
+		}
 	} catch (error) {
 		console.log(error);
 		res.status(500).json({ success: false, error: 'Something went wrong' });
@@ -64,8 +74,18 @@ router.put('/:id', async (req, res) => {
 //Delete an idea
 router.delete('/:id', async (req, res) => {
 	try {
-		await Idea.findByIdAndDelete(req.params.id);
-		res.json({ success: true, data: {} });
+		const idea = await Idea.findById(req.params.id);
+
+		//Match the usernames
+		if (idea.username === req.body.username) {
+			await Idea.findByIdAndDelete(req.params.id);
+			res.json({ success: true, data: {} });
+		}
+
+		//Usernames don't match
+		else {
+			res.status(403).json({ success: false, error: 'Unauthorized to delete' });
+		}
 	} catch (error) {
 		console.log(error);
 		res.status(500).json({ success: false, error: 'Something went wrong' });
